+++ .storybook/preview.tsx
@@ -1,8 +1,8 @@
 import { withThemeByClassName } from '@storybook/addon-themes';
-import { INITIAL_VIEWPORTS, MINIMAL_VIEWPORTS } from '@storybook/addon-viewport';
-import type { Preview } from '@storybook/react';
+import type { Preview } from '@storybook/react-vite';
+import { INITIAL_VIEWPORTS, MINIMAL_VIEWPORTS } from 'storybook/viewport';
 
-import '../src/styles/globals.css';
+import '../src/index.css';
 
 // Custom viewports
 const CUSTOM_VIEWPORTS = {
@@ -52,7 +52,6 @@ const CUSTOM_VIEWPORTS = {
 
 const preview: Preview = {
 	parameters: {
-		actions: { argTypesRegex: '^on[A-Z].*' },
 		controls: {
 			matchers: {
 				color: /(background|color)$/i,
@@ -69,6 +68,7 @@ const preview: Preview = {
 		// Default configuration
 		layout: 'centered',
 	},
+
 	decorators: [
 		withThemeByClassName({
 			themes: {
@@ -78,6 +78,8 @@ const preview: Preview = {
 			defaultTheme: 'light',
 		}),
 	],
+
+	tags: ['autodocs'],
 };
 
 export default preview;